+++ src/main/scala/org/podval/tools/scalajs/ModuleKind.scala
@@ -1,4 +1,4 @@
-package org.podval.tools.scalajs.js
+package org.podval.tools.scalajs
 
 enum ModuleKind derives CanEqual:
   case NoModule