+++ src/main/scala/org/podval/tools/scalajs/ModuleSplitStyle.scala
@@ -1,4 +1,4 @@
-package org.podval.tools.scalajs.js
+package org.podval.tools.scalajs
 
 enum ModuleSplitStyle derives CanEqual:
   case FewestModules