+++ src/main/scala/org/podval/tools/scalajs/Optimization.scala
@@ -1,4 +1,4 @@
-package org.podval.tools.scalajs.js
+package org.podval.tools.scalajs
 
 enum Optimization(val description: String) derives CanEqual:
   case Fast extends Optimization(description = " - fast")