+++ src/main/scala/org/podval/tools/scalajsplugin/ScalaJSPlugin.scala
@@ -1,11 +1,9 @@
-package org.podval.tools.scalajs
+package org.podval.tools.scalajsplugin
 
 import org.gradle.api.plugins.JavaPlugin
 import org.gradle.api.plugins.scala.ScalaBasePlugin
 import org.gradle.api.{Plugin, Project}
 import org.podval.tools.build.{Gradle, GradleClassPath, ScalaLibrary}
-import org.podval.tools.scalajs.js.JSDelegate
-import org.podval.tools.scalajs.jvm.JvmDelegate
 import scala.jdk.CollectionConverters.IterableHasAsScala
 
 final class ScalaJSPlugin extends Plugin[Project]:
@@ -16,25 +14,25 @@ final class ScalaJSPlugin extends Plugin[Project]:
       Option(project.findProperty(ScalaJSPlugin.maiflaiProperty )).isDefined ||
       Option(project.findProperty(ScalaJSPlugin.disabledProperty)).exists(_.toString.toBoolean)
 
-    val delegate: ScalaJSPlugin.Delegate = if isJSDisabled then JvmDelegate() else JSDelegate()
+    val delegate: ScalaJSPlugin.Delegate = if isJSDisabled then JvmDelegate() else ScalaJSDelegate()
 
     delegate.beforeEvaluate(project)
-    
+
     project.afterEvaluate((project: Project) =>
       val pluginScalaLibrary: ScalaLibrary = ScalaLibrary.getFromClasspath(GradleClassPath.collect(this))
       val projectScalaLibrary: ScalaLibrary = ScalaLibrary.getFromConfiguration(Gradle.getConfiguration(project, JavaPlugin.IMPLEMENTATION_CONFIGURATION_NAME))
-  
+
       // AnalysisDetector, which runs during execution of TestTask, needs Zinc classes;
       // if I ever get rid of it, this classpath expansion goes away.
       // TODO instead, add configuration itself to whatever configuration lists dependencies available to the plugin... "classpath"?
       GradleClassPath.addTo(this, Gradle.getConfiguration(project, ScalaBasePlugin.ZINC_CONFIGURATION_NAME).asScala)
-  
+
       delegate.afterEvaluate(
         project,
         pluginScalaLibrary,
         projectScalaLibrary
       )
-  
+
       projectScalaLibrary.verify(
         ScalaLibrary.getFromClasspath(Gradle.getConfiguration(project, JavaPlugin.RUNTIME_CLASSPATH_CONFIGURATION_NAME).asScala)
       )
@@ -46,7 +44,7 @@ object ScalaJSPlugin:
 
   abstract class Delegate:
     def beforeEvaluate(project: Project): Unit
-    
+
     def afterEvaluate(
       project: Project,
       pluginScalaLibrary : ScalaLibrary,